[op][n_upsable][appshell] Add injection API for clearing all database of sitefilter feature
:Release Notes:
Add injection API for clearing all database of sitefilter feature

:Detailed Notes:
This commit will add support for the 'Reset Default Settings' feature.
Currently, we only clear URL data in the current mode
(off, approve, block mode). With the 'Reset Default Settings' feature,
 all URLs in those modes must be cleared at once

:Testing Performed:
1. Run Enact browser Application
2. Tap to "Menu" icon, then select "Settings"
3. Tap to "Site Filtering"
4. When Pin popup is opended, entering PIN "0000" => tap to "OK" button
5. Tap to "Approved Sites" => "Add" icon
6. Input list of url, such as: youtube.com, google.com ...
7. Tap to "Blocked Sites" => "Add" icon
8. Input list of url, such as: dantri.com, 24h.com ...
9. Back to "Setting" menu
10. tap to "Reset Settings" button => "Yes" button
11. Open "Site Filtering" feature again, all url in "Approved Sites"
and "Blocked Sites" must be deleted.

:QA Notes:
Test-cases: QWTC-13405

:Issues Addressed:
[WRP-27871] Add injection API for clearing all database
of sitefilter feature

Reviewed-on: http://gpro.lge.com/c/webosose/chromium108/+/375846
Tested-by: Van Luc Tran <luc2.tran@lge.com>

+++ src/neva/browser_service/browser/sitefilter_service_impl.cc
@@ -132,23 +132,50 @@ void SiteFilterServiceImpl::AddURL(const std::string& url,
 }
 
 void SiteFilterServiceImpl::DeleteURLs(const std::vector<std::string>& urls,
+                                       bool is_delete_all,
                                        DeleteURLsCallback callback) {
-  if (type_ == Type::kDisabled) {
-    LOG(ERROR) << __func__ << " Site Filter type is OFF !";
-    std::move(callback).Run(false);
-    return;
-  }
+  if (!is_delete_all) {
+    if (type_ == Type::kDisabled) {
+      LOG(ERROR) << __func__ << " Site Filter type is OFF !";
+      std::move(callback).Run(false);
+      return;
+    }
 
-  if (!url_list_table_->DeleteURLs(urls)) {
-    LOG(ERROR) << __func__ << " Unable to Remove URLs from DB";
-    std::move(callback).Run(false);
-    return;
-  }
+    if (!url_list_table_->DeleteURLs(urls)) {
+      LOG(ERROR) << __func__ << " Unable to Remove URLs from DB";
+      std::move(callback).Run(false);
+      return;
+    }
 
-  for (const auto& url : urls) {
-    url_list_.erase(url);
+    for (const auto& url : urls) {
+      url_list_.erase(url);
+    }
+  } else {
+    if (url_list_table_) {
+      std::unique_ptr<URLDatabase> remaining_database(new URLDatabase(
+          GetTableName() == kAllowURLTableName ? kBlockURLTableName
+                                               : kAllowURLTableName));
+      if (!url_list_table_->ClearAllURLs() ||
+          !remaining_database->ClearAllURLs()) {
+        LOG(ERROR) << __func__
+                   << " Unable to Remove All URLs from Block and Allow DB";
+        std::move(callback).Run(false);
+        return;
+      }
+    } else {
+      std::unique_ptr<URLDatabase> allow_database(
+          new URLDatabase(kAllowURLTableName));
+      std::unique_ptr<URLDatabase> block_database(
+          new URLDatabase(kBlockURLTableName));
+      if (!allow_database->ClearAllURLs() || !block_database->ClearAllURLs()) {
+        LOG(ERROR) << __func__
+                   << " Unable to Remove All URLs from Block and Allow DB";
+        std::move(callback).Run(false);
+        return;
+      }
+    }
+    url_list_.clear();
   }
-
   std::move(callback).Run(true);
 }
